+++ HISTORY.md
@@ -4,12 +4,18 @@ All notable changes to this project will be documented in this file.
 
 ## unreleased
 
+* Added
+  * Class `Models.ExternalReference` got a new property `hashes` ([#984] via [#985])
+  * Serializers and `ExternalReference`-Normalizers will take `Models.ExternalReference.hashes` into account ([#984] via [#985])
+
 * Build
   * Use _Webpack_ `v5.89.0` now, was `v5.88.2` (via [#979])
   * Use _ts-loader_ `v9.5.0` now, was `v9.4.4` (via [#977])
 
 [#977]: https://github.com/CycloneDX/cyclonedx-javascript-library/pull/977
 [#979]: https://github.com/CycloneDX/cyclonedx-javascript-library/pull/979 
+[#984]: https://github.com/CycloneDX/cyclonedx-javascript-library/issues/984
+[#985]: https://github.com/CycloneDX/cyclonedx-javascript-library/pull/985
 
 ## 6.0.0 -- 2023-08-26